+++ src/test/kotlin/com/group/mock/place/TestPlaceRepository.kt
@@ -1,0 +1,8 @@
+package com.group.mock.place
+
+import com.group.mock.place.domain.Place
+import com.group.mock.place.infrastructure.PlaceRepository
+import com.group.mock.infrastructure.jpa.TestJpaRepository
+
+class TestPlaceRepository : TestJpaRepository<Place, Long>("id"), PlaceRepository {
+}
