--- src/test/kotlin/com/group/mock/v1/food/TestFoodRepositoryV1.kt
@@ -1,8 +1,0 @@
-package com.group.mock.v1.food
-
-import com.group.mock.food.domain.Food
-import com.group.mock.food.infrastructure.FoodRepository
-import com.group.mock.v1.infrastructure.jpa.TestJpaRepositoryV1
-
-class TestFoodRepositoryV1 : TestJpaRepositoryV1<Food, Long>("id"), FoodRepository {
-}
