--- src/test/kotlin/com/group/mock/v1/place/TestPlaceRepositoryV1.kt
@@ -1,8 +1,0 @@
-package com.group.mock.v1.place
-
-import com.group.mock.place.domain.Place
-import com.group.mock.place.infrastructure.PlaceRepository
-import com.group.mock.v1.infrastructure.jpa.TestJpaRepositoryV1
-
-class TestPlaceRepositoryV1 : TestJpaRepositoryV1<Place, Long>("id"), PlaceRepository {
-}
