--- src/test/kotlin/com/group/mock/v2/food/TestFoodRepositoryV2.kt
@@ -1,9 +1,0 @@
-package com.group.mock.v2.food
-
-import com.group.mock.food.domain.Food
-import com.group.mock.food.infrastructure.FoodRepository
-import com.group.mock.v2.common.FOOD_TEST_DB
-import com.group.mock.v2.common.jpa.TestJpaRepositoryV2
-
-class TestFoodRepositoryV2 : TestJpaRepositoryV2<Food, Long>(FOOD_TEST_DB), FoodRepository {
-}
