--- src/test/kotlin/com/group/mock/v2/place/TestPlaceRepositoryV2.kt
@@ -1,9 +1,0 @@
-package com.group.mock.v2.place
-
-import com.group.mock.place.domain.Place
-import com.group.mock.place.infrastructure.PlaceRepository
-import com.group.mock.v2.common.PLACE_TEST_DB
-import com.group.mock.v2.common.jpa.TestJpaRepositoryV2
-
-class TestPlaceRepositoryV2 : TestJpaRepositoryV2<Place, Long>(PLACE_TEST_DB), PlaceRepository {
-}
